ATLAS-5317: Restore pre-5317 purge test layout in AdminPurgeTest, PurgeServiceTest, and AdminResourceTest.
